Revert "feat: Add MomentWrite request and response for API"
This reverts commit c878857596acd2c368f5087c459da2c7e95fc8fa.
