+++ frontend/Moment/app/src/main/java/snu/swpp/moment/api/LoginRequest.java
@@ -1,21 +1,16 @@
-package snu.swpp.moment.api.request;
+package snu.swpp.moment.api;
 
 import com.google.gson.annotations.SerializedName;
 
-public class RegisterRequest {
+public class LoginRequest {
 
     @SerializedName("username")
     private String username;
-
     @SerializedName("password")
     private String password;
 
-    @SerializedName("nickname")
-    private String nickname;
-
-    public RegisterRequest(String username, String password, String nickname) {
+    public LoginRequest(String username, String password) {
         this.username = username;
         this.password = password;
-        this.nickname = nickname;
     }
 }
