+++ frontend/Moment/app/src/main/java/snu/swpp/moment/api/RegisterRequest.java
@@ -1,16 +1,21 @@
-package snu.swpp.moment.api.request;
+package snu.swpp.moment.api;
 
 import com.google.gson.annotations.SerializedName;
 
-public class LoginRequest {
+public class RegisterRequest {
 
     @SerializedName("username")
     private String username;
+
     @SerializedName("password")
     private String password;
 
-    public LoginRequest(String username, String password) {
+    @SerializedName("nickname")
+    private String nickname;
+
+    public RegisterRequest(String username, String password, String nickname) {
         this.username = username;
         this.password = password;
+        this.nickname = nickname;
     }
 }
