+++ frontend/Moment/app/src/main/java/snu/swpp/moment/api/TokenVerifyRequest.java
@@ -1,4 +1,4 @@
-package snu.swpp.moment.api.request;
+package snu.swpp.moment.api;
 
 import com.google.gson.annotations.SerializedName;
