--- frontend/Moment/app/src/main/java/snu/swpp/moment/api/response/MomentPostResponse.java
@@ -1,5 +1,0 @@
-package snu.swpp.moment.api.response;
-
-public class MomentPostResponse {
-
-}
