--- frontend/Moment/app/src/main/java/snu/swpp/moment/data/model/MomentPair.java
@@ -1,24 +1,0 @@
-package snu.swpp.moment.data.model;
-
-import java.util.Date;
-
-
-public class MomentPair {
-    private long id;
-    private String moment;
-    private String reply;
-    private Date moment_created_at;
-    private Date reply_created_at;
-
-    public MomentPair(long id, String moment, String reply, long moment_created_at, long reply_created_at) {
-        this.id = id;
-        this.moment = moment;
-        this.reply = reply;
-        this.moment_created_at = convertTimestampToDate(moment_created_at);
-        this.reply_created_at = convertTimestampToDate(reply_created_at);
-    }
-
-    private Date convertTimestampToDate(long timestamp) {
-        return new Date(timestamp);
-    }
-}
